Fix shift key combo key press
Shift+X when passed to press() and used in onjunction with down() and
up() now will change the casing of the character, as long as the
character is in the format `KeyX`. A key press that is `x` or `X` will
ignore the shift modifier key. This fix matches the playwright
behaviour.
Fixes: #285