Fix keyboard symbol input
Some tests were failing due to the new fix for #285. To resolve this we
needed to ignore the shift key for anything other than a `KeyX` (where
X is from A-Z) and to use the shift key for keys where the shift key
needs to be used.

Another edge case was found when we passed in `+++`, which should type
out `++`, but `++` should panic since this should be split into ["+",
""].

Resolves: https://github.com/grafana/xk6-browser/pull/326/commits/2d77a393c9708449d2c6a3ac3ce6f2f79d1235bf#r910097593
Resolves: https://github.com/grafana/xk6-browser/pull/326/commits/2d77a393c9708449d2c6a3ac3ce6f2f79d1235bf#r910103653